+++ src/modules/history/defaultState.js
@@ -1,0 +1,8 @@
+export const getDefaultState = () => {
+  return {
+    days: {},
+    fetching: false,
+    error: false,
+    baseId: ''
+  };
+};
