+++ src/modules/user/defaultState.js
@@ -1,0 +1,8 @@
+export const getDefaultState = () => {
+  return {
+    account: null,
+    balances: {},
+    fetching: false,
+    error: false
+  };
+};
